working aside from custom headers not getting through
